// sdks/typescript/core-generated/runtime-protobuf/runtime/v1/runtime_target_identity.ts
export interface RuntimeDurableLocalTargetRef {
    /**
     * @generated from protobuf field: string version = 1
     */
    version: string;
    /**
     * @generated from protobuf oneof: ref
     */
    ref: {
        oneofKind: "profileBindingId";
        /**
         * @generated from protobuf field: string profile_binding_id = 2
         */
        profileBindingId: string;
    } | {
        oneofKind: "readinessRef";
        /**
         * @generated from protobuf field: string readiness_ref = 3
         */
        readinessRef: string;
    } | {
        oneofKind: undefined;
    };
}
/**
 * @generated from protobuf message nimi.runtime.v1.RuntimeDurableCloudTargetRef
 */
export interface RuntimeDurableCloudTargetRef {
    /**
     * @generated from protobuf field: string version = 1
     */
    version: string;
    /**
     * @generated from protobuf field: string connector_id = 2
     */
    connectorId: string;
    /**
     * @generated from protobuf field: string remote_model_catalog_id = 3
     */
    remoteModelCatalogId: string;
    /**
     * @generated from protobuf field: string provider_model_id = 4
     */
    providerModelId: string;
    /**
     * @generated from protobuf field: string provider = 5
     */
    provider: string;
}
/**
 * @generated from protobuf message nimi.runtime.v1.RuntimeDurableTargetRef
 */
export interface RuntimeDurableTargetRef {
    /**
     * @generated from protobuf oneof: target
     */
    target: {
        oneofKind: "localRuntime";
        /**
         * @generated from protobuf field: nimi.runtime.v1.RuntimeDurableLocalTargetRef local_runtime = 1
         */
        localRuntime: RuntimeDurableLocalTargetRef;
    } | {
        oneofKind: "cloud";
        /**
         * @generated from protobuf field: nimi.runtime.v1.RuntimeDurableCloudTargetRef cloud = 2
         */
        cloud: RuntimeDurableCloudTargetRef;
    } | {
        oneofKind: undefined;
    };
}
// ...
class RuntimeDurableLocalTargetRef$Type extends MessageType<RuntimeDurableLocalTargetRef> {
    constructor() {
        super("nimi.runtime.v1.RuntimeDurableLocalTargetRef", [
            { no: 1, name: "version", kind: "scalar", T: 9 /*ScalarType.STRING*/ },
            { no: 2, name: "profile_binding_id", kind: "scalar", oneof: "ref", T: 9 /*ScalarType.STRING*/ },
            { no: 3, name: "readiness_ref", kind: "scalar", oneof: "ref", T: 9 /*ScalarType.STRING*/ }
        ]);
    }
    create(value?: PartialMessage<RuntimeDurableLocalTargetRef>): RuntimeDurableLocalTargetRef {
        const message = globalThis.Object.create((this.messagePrototype!));
        message.version = "";
        message.ref = { oneofKind: undefined };
        if (value !== undefined)
            reflectionMergePartial<RuntimeDurableLocalTargetRef>(this, message, value);
        return message;
    }
    internalBinaryRead(reader: IBinaryReader, length: number, options: BinaryReadOptions, target?: RuntimeDurableLocalTargetRef): RuntimeDurableLocalTargetRef {
        let message = target ?? this.create(), end = reader.pos + length;
        while (reader.pos < end) {
            let [fieldNo, wireType] = reader.tag();
            switch (fieldNo) {
                case /* string version */ 1:
                    message.version = reader.string();
                    break;
                case /* string profile_binding_id */ 2:
                    message.ref = {
                        oneofKind: "profileBindingId",
                        profileBindingId: reader.string()
                    };
                    break;
                case /* string readiness_ref */ 3:
                    message.ref = {
                        oneofKind: "readinessRef",
                        readinessRef: reader.string()
                    };
                    break;
                default:
                    let u = options.readUnknownField;
                    if (u === "throw")
                        throw new globalThis.Error(`Unknown field ${fieldNo} (wire type ${wireType}) for ${this.typeName}`);
                    let d = reader.skip(wireType);
                    if (u !== false)
                        (u === true ? UnknownFieldHandler.onRead : u)(this.typeName, message, fieldNo, wireType, d);
            }
        }
        return message;
    }
    internalBinaryWrite(message: RuntimeDurableLocalTargetRef, writer: IBinaryWriter, options: BinaryWriteOptions): IBinaryWriter {
        /* string version = 1; */
        if (message.version !== "")
            writer.tag(1, WireType.LengthDelimited).string(message.version);
        /* string profile_binding_id = 2; */
        if (message.ref.oneofKind === "profileBindingId")
            writer.tag(2, WireType.LengthDelimited).string(message.ref.profileBindingId);
        /* string readiness_ref = 3; */
        if (message.ref.oneofKind === "readinessRef")
            writer.tag(3, WireType.LengthDelimited).string(message.ref.readinessRef);
        let u = options.writeUnknownFields;
        if (u !== false)
            (u == true ? UnknownFieldHandler.onWrite : u)(this.typeName, message, writer);
        return writer;
    }
}
/**
 * @generated MessageType for protobuf message nimi.runtime.v1.RuntimeDurableLocalTargetRef
 */
export const RuntimeDurableLocalTargetRef = new RuntimeDurableLocalTargetRef$Type();
// ...
export const RuntimeDurableCloudTargetRef = new RuntimeDurableCloudTargetRef$Type();
// @generated message type with reflection information, may provide speed optimized methods
class RuntimeDurableTargetRef$Type extends MessageType<RuntimeDurableTargetRef> {
    constructor() {
        super("nimi.runtime.v1.RuntimeDurableTargetRef", [
            { no: 1, name: "local_runtime", kind: "message", oneof: "target", T: () => RuntimeDurableLocalTargetRef },
            { no: 2, name: "cloud", kind: "message", oneof: "target", T: () => RuntimeDurableCloudTargetRef }
        ]);
    }
    create(value?: PartialMessage<RuntimeDurableTargetRef>): RuntimeDurableTargetRef {
        const message = globalThis.Object.create((this.messagePrototype!));
        message.target = { oneofKind: undefined };
        if (value !== undefined)
            reflectionMergePartial<RuntimeDurableTargetRef>(this, message, value);
        return message;
    }
    internalBinaryRead(reader: IBinaryReader, length: number, options: BinaryReadOptions, target?: RuntimeDurableTargetRef): RuntimeDurableTargetRef {
        let message = target ?? this.create(), end = reader.pos + length;
        while (reader.pos < end) {
            let [fieldNo, wireType] = reader.tag();
            switch (fieldNo) {
                case /* nimi.runtime.v1.RuntimeDurableLocalTargetRef local_runtime */ 1:
                    message.target = {
                        oneofKind: "localRuntime",
                        localRuntime: RuntimeDurableLocalTargetRef.internalBinaryRead(reader, reader.uint32(), options, (message.target as any).localRuntime)
                    };
                    break;
                case /* nimi.runtime.v1.RuntimeDurableCloudTargetRef cloud */ 2:
                    message.target = {
                        oneofKind: "cloud",
                        cloud: RuntimeDurableCloudTargetRef.internalBinaryRead(reader, reader.uint32(), options, (message.target as any).cloud)
                    };
                    break;
                default:
                    let u = options.readUnknownField;
                    if (u === "throw")
                        throw new globalThis.Error(`Unknown field ${fieldNo} (wire type ${wireType}) for ${this.typeName}`);
                    let d = reader.skip(wireType);
                    if (u !== false)
                        (u === true ? UnknownFieldHandler.onRead : u)(this.typeName, message, fieldNo, wireType, d);
            }
        }
        return message;
    }
    internalBinaryWrite(message: RuntimeDurableTargetRef, writer: IBinaryWriter, options: BinaryWriteOptions): IBinaryWriter {
        /* nimi.runtime.v1.RuntimeDurableLocalTargetRef local_runtime = 1; */
        if (message.target.oneofKind === "localRuntime")
            RuntimeDurableLocalTargetRef.internalBinaryWrite(message.target.localRuntime, writer.tag(1, WireType.LengthDelimited).fork(), options).join();
        /* nimi.runtime.v1.RuntimeDurableCloudTargetRef cloud = 2; */
        if (message.target.oneofKind === "cloud")
            RuntimeDurableCloudTargetRef.internalBinaryWrite(message.target.cloud, writer.tag(2, WireType.LengthDelimited).fork(), options).join();
        let u = options.writeUnknownFields;
        if (u !== false)
            (u == true ? UnknownFieldHandler.onWrite : u)(this.typeName, message, writer);
        return writer;
    }
}
/**
 * @generated MessageType for protobuf message nimi.runtime.v1.RuntimeDurableTargetRef
 */
export const RuntimeDurableTargetRef = new RuntimeDurableTargetRef$Type();